Separate out `verifier_api` module in subspace-fraud-proof
