+++ crates/subspace-fraud-proof/src/invalid_state_transition_proof.rs
@@ -6,17 +6,17 @@
 //! specific extrinsic execution are supported.
 
 use crate::domain_extrinsics_builder::BuildDomainExtrinsics;
+use crate::verifier_api::VerifierApi;
 use codec::{Codec, Decode, Encode};
 use domain_runtime_primitives::opaque::Block;
 use hash_db::{HashDB, Hasher, Prefix};
-use sc_client_api::{backend, HeaderBackend};
+use sc_client_api::backend;
 use sp_api::{ProvideRuntimeApi, StorageProof};
 use sp_core::traits::{CodeExecutor, RuntimeCode, SpawnNamed};
 use sp_core::H256;
 use sp_domain_digests::AsPredigest;
 use sp_domains::fraud_proof::{ExecutionPhase, InvalidStateTransitionProof, VerificationError};
-use sp_domains::{DomainId, ExecutorApi};
-use sp_receipts::ReceiptsApi;
+use sp_domains::ExecutorApi;
 use sp_runtime::traits::{BlakeTwo256, Block as BlockT, HashFor, Header as HeaderT, NumberFor};
 use sp_runtime::{Digest, DigestItem};
 use sp_state_machine::backend::AsTrieBackend;
@@ -193,31 +193,31 @@ pub struct InvalidStateTransitionProofVerifier<
     Exec,
     Spawn,
     Hash,
-    PrePostStateRootVerifier,
+    VerifierClient,
     DomainExtrinsicsBuilder,
 > {
     client: Arc<C>,
     executor: Exec,
     spawn_handle: Spawn,
-    pre_post_state_root_verifier: PrePostStateRootVerifier,
+    verifier_client: VerifierClient,
     domain_extrinsics_builder: DomainExtrinsicsBuilder,
     _phantom: PhantomData<(PBlock, Hash)>,
 }
 
-impl<PBlock, C, Exec, Spawn, Hash, PrePostStateRootVerifier, DomainExtrinsicsBuilder> Clone
+impl<PBlock, C, Exec, Spawn, Hash, VerifierClient, DomainExtrinsicsBuilder> Clone
     for InvalidStateTransitionProofVerifier<
         PBlock,
         C,
         Exec,
         Spawn,
         Hash,
-        PrePostStateRootVerifier,
+        VerifierClient,
         DomainExtrinsicsBuilder,
     >
 where
     Exec: Clone,
     Spawn: Clone,
-    PrePostStateRootVerifier: Clone,
+    VerifierClient: Clone,
     DomainExtrinsicsBuilder: Clone,
 {
     fn clone(&self) -> Self {
@@ -225,183 +225,21 @@ where
             client: self.client.clone(),
             executor: self.executor.clone(),
             spawn_handle: self.spawn_handle.clone(),
-            pre_post_state_root_verifier: self.pre_post_state_root_verifier.clone(),
+            verifier_client: self.verifier_client.clone(),
             domain_extrinsics_builder: self.domain_extrinsics_builder.clone(),
             _phantom: self._phantom,
         }
     }
 }
 
-/// Verifies `pre_state_root` and `post_state_root` in [`InvalidStateTransitionProof`].
-pub trait VerifyPrePostStateRoot {
-    /// Verifies whether `pre_state_root` declared in the proof is same as the one recorded on chain.
-    fn verify_pre_state_root(
-        &self,
-        invalid_state_transition_proof: &InvalidStateTransitionProof,
-    ) -> Result<(), VerificationError>;
-
-    /// Verifies whether `post_state_root` declared in the proof is different from the one recorded on chain.
-    fn verify_post_state_root(
-        &self,
-        invalid_state_transition_proof: &InvalidStateTransitionProof,
-    ) -> Result<(), VerificationError>;
-
-    /// Returns the hash of primary block at height `domain_block_number`.
-    fn primary_hash(
-        &self,
-        domain_id: DomainId,
-        domain_block_number: u32,
-    ) -> Result<H256, VerificationError>;
-}
-
-/// Verifier of `pre_state_root` in [`InvalidStateTransitionProof`].
-///
-/// `client` could be either the primary chain client or system domain client.
-pub struct PrePostStateRootVerifier<Client, Block> {
-    client: Arc<Client>,
-    _phantom: PhantomData<Block>,
-}
-
-impl<Client, Block> Clone for PrePostStateRootVerifier<Client, Block> {
-    fn clone(&self) -> Self {
-        Self {
-            client: self.client.clone(),
-            _phantom: self._phantom,
-        }
-    }
-}
-
-impl<Client, Block> PrePostStateRootVerifier<Client, Block> {
-    /// Constructs a new instance of [`PrePostStateRootVerifier`].
-    pub fn new(client: Arc<Client>) -> Self {
-        Self {
-            client,
-            _phantom: Default::default(),
-        }
-    }
-}
-
-impl<Client, Block> VerifyPrePostStateRoot for PrePostStateRootVerifier<Client, Block>
-where
-    Block: BlockT,
-    Client: ProvideRuntimeApi<Block> + HeaderBackend<Block>,
-    Client::Api: ReceiptsApi<Block, domain_runtime_primitives::Hash>,
-{
-    // TODO: It's not necessary to require `pre_state_root` in the proof and then verify, it can
-    // be just retrieved by the verifier itself according the execution phase, which requires some
-    // fixes in tests however, we can do this refactoring once we have or are able to construct a
-    // proper `VerifyPrePostStateRoot` implementation in test.
-    //
-    // Related: https://github.com/subspace/subspace/pull/1240#issuecomment-1476212007
-    fn verify_pre_state_root(
-        &self,
-        invalid_state_transition_proof: &InvalidStateTransitionProof,
-    ) -> Result<(), VerificationError> {
-        let InvalidStateTransitionProof {
-            domain_id,
-            parent_number,
-            bad_receipt_hash,
-            pre_state_root,
-            execution_phase,
-            ..
-        } = invalid_state_transition_proof;
-
-        let pre_state_root_onchain = match execution_phase {
-            ExecutionPhase::InitializeBlock { domain_parent_hash } => {
-                self.client.runtime_api().state_root(
-                    self.client.info().best_hash,
-                    *domain_id,
-                    NumberFor::<Block>::from(*parent_number),
-                    Block::Hash::decode(&mut domain_parent_hash.encode().as_slice())?,
-                )?
-            }
-            ExecutionPhase::ApplyExtrinsic(trace_index_of_pre_state_root)
-            | ExecutionPhase::FinalizeBlock {
-                total_extrinsics: trace_index_of_pre_state_root,
-            } => {
-                let trace = self.client.runtime_api().execution_trace(
-                    self.client.info().best_hash,
-                    *domain_id,
-                    *bad_receipt_hash,
-                )?;
-
-                trace.get(*trace_index_of_pre_state_root as usize).copied()
-            }
-        };
-
-        match pre_state_root_onchain {
-            Some(expected_pre_state_root) if expected_pre_state_root == *pre_state_root => Ok(()),
-            res => {
-                tracing::debug!(
-                    "Invalid `pre_state_root` in InvalidStateTransitionProof for {domain_id:?}, expected: {res:?}, got: {pre_state_root:?}",
-                );
-                Err(VerificationError::InvalidPreStateRoot)
-            }
-        }
-    }
-
-    fn verify_post_state_root(
-        &self,
-        invalid_state_transition_proof: &InvalidStateTransitionProof,
-    ) -> Result<(), VerificationError> {
-        let InvalidStateTransitionProof {
-            domain_id,
-            bad_receipt_hash,
-            execution_phase,
-            post_state_root,
-            ..
-        } = invalid_state_transition_proof;
-
-        let trace = self.client.runtime_api().execution_trace(
-            self.client.info().best_hash,
-            *domain_id,
-            *bad_receipt_hash,
-        )?;
-
-        let post_state_root_onchain = match execution_phase {
-            ExecutionPhase::InitializeBlock { .. } => trace
-                .get(0)
-                .ok_or(VerificationError::PostStateRootNotFound)?,
-            ExecutionPhase::ApplyExtrinsic(trace_index_of_post_state_root)
-            | ExecutionPhase::FinalizeBlock {
-                total_extrinsics: trace_index_of_post_state_root,
-            } => trace
-                .get(*trace_index_of_post_state_root as usize + 1)
-                .ok_or(VerificationError::PostStateRootNotFound)?,
-        };
-
-        if post_state_root_onchain == post_state_root {
-            Err(VerificationError::SamePostStateRoot)
-        } else {
-            Ok(())
-        }
-    }
-
-    fn primary_hash(
-        &self,
-        domain_id: DomainId,
-        domain_block_number: u32,
-    ) -> Result<H256, VerificationError> {
-        self.client
-            .runtime_api()
-            .primary_hash(
-                self.client.info().best_hash,
-                domain_id,
-                domain_block_number.into(),
-            )?
-            .and_then(|primary_hash| Decode::decode(&mut primary_hash.encode().as_slice()).ok())
-            .ok_or(VerificationError::PrimaryHashNotFound)
-    }
-}
-
-impl<PBlock, C, Exec, Spawn, Hash, PrePostStateRootVerifier, DomainExtrinsicsBuilder>
+impl<PBlock, C, Exec, Spawn, Hash, VerifierClient, DomainExtrinsicsBuilder>
     InvalidStateTransitionProofVerifier<
         PBlock,
         C,
         Exec,
         Spawn,
         Hash,
-        PrePostStateRootVerifier,
+        VerifierClient,
         DomainExtrinsicsBuilder,
     >
 where
@@ -412,7 +250,7 @@ where
     Exec: CodeExecutor + Clone + 'static,
     Spawn: SpawnNamed + Clone + Send + 'static,
     Hash: Encode + Decode,
-    PrePostStateRootVerifier: VerifyPrePostStateRoot,
+    VerifierClient: VerifierApi,
     DomainExtrinsicsBuilder: BuildDomainExtrinsics<PBlock>,
 {
     /// Constructs a new instance of [`InvalidStateTransitionProofVerifier`].
@@ -420,14 +258,14 @@ where
         client: Arc<C>,
         executor: Exec,
         spawn_handle: Spawn,
-        pre_post_state_root_verifier: PrePostStateRootVerifier,
+        verifier_client: VerifierClient,
         domain_extrinsics_builder: DomainExtrinsicsBuilder,
     ) -> Self {
         Self {
             client,
             executor,
             spawn_handle,
-            pre_post_state_root_verifier,
+            verifier_client,
             domain_extrinsics_builder,
             _phantom: PhantomData::<(PBlock, Hash)>,
         }
@@ -438,10 +276,10 @@ where
         &self,
         invalid_state_transition_proof: &InvalidStateTransitionProof,
     ) -> Result<(), VerificationError> {
-        self.pre_post_state_root_verifier
+        self.verifier_client
             .verify_pre_state_root(invalid_state_transition_proof)?;
 
-        self.pre_post_state_root_verifier
+        self.verifier_client
             .verify_post_state_root(invalid_state_transition_proof)?;
 
         let InvalidStateTransitionProof {
@@ -478,7 +316,7 @@ where
                 let primary_number = parent_number + 1;
                 let digest = if domain_id.is_system() {
                     let primary_hash = self
-                        .pre_post_state_root_verifier
+                        .verifier_client
                         .primary_hash(*domain_id, primary_number)?;
                     Digest {
                         logs: vec![DigestItem::primary_block_info::<NumberFor<Block>, _>((
@@ -500,7 +338,7 @@ where
             }
             ExecutionPhase::ApplyExtrinsic(extrinsic_index) => {
                 let primary_hash = self
-                    .pre_post_state_root_verifier
+                    .verifier_client
                     .primary_hash(*domain_id, parent_number + 1)?;
                 let domain_extrinsics = self
                     .domain_extrinsics_builder
@@ -554,7 +392,7 @@ pub trait VerifyInvalidStateTransitionProof {
     ) -> Result<(), VerificationError>;
 }
 
-impl<PBlock, C, Exec, Spawn, Hash, PrePostStateRootVerifier, DomainExtrinsicsBuilder>
+impl<PBlock, C, Exec, Spawn, Hash, VerifierClient, DomainExtrinsicsBuilder>
     VerifyInvalidStateTransitionProof
     for InvalidStateTransitionProofVerifier<
         PBlock,
@@ -562,7 +400,7 @@ impl<PBlock, C, Exec, Spawn, Hash, PrePostStateRootVerifier, DomainExtrinsicsBui
         Exec,
         Spawn,
         Hash,
-        PrePostStateRootVerifier,
+        VerifierClient,
         DomainExtrinsicsBuilder,
     >
 where
@@ -573,7 +411,7 @@ where
     Exec: CodeExecutor + Clone + 'static,
     Spawn: SpawnNamed + Clone + Send + 'static,
     Hash: Encode + Decode,
-    PrePostStateRootVerifier: VerifyPrePostStateRoot,
+    VerifierClient: VerifierApi,
     DomainExtrinsicsBuilder: BuildDomainExtrinsics<PBlock>,
 {
     fn verify_invalid_state_transition_proof(
