+++ crates/subspace-service/src/lib.rs
@@ -77,7 +77,7 @@ use std::num::NonZeroUsize;
 use std::sync::{Arc, Mutex};
 use subspace_core_primitives::crypto::kzg::{embedded_kzg_settings, Kzg};
 use subspace_fraud_proof::domain_extrinsics_builder::SystemDomainExtrinsicsBuilder;
-use subspace_fraud_proof::invalid_state_transition_proof::PrePostStateRootVerifier;
+use subspace_fraud_proof::verifier_api::VerifierClient;
 use subspace_networking::libp2p::multiaddr::Protocol;
 use subspace_networking::libp2p::Multiaddr;
 use subspace_networking::utils::online_status_informer;
@@ -138,7 +138,7 @@ pub type InvalidStateTransitionProofVerifier<RuntimeApi, ExecutorDispatch> =
         NativeElseWasmExecutor<ExecutorDispatch>,
         SpawnTaskHandle,
         Hash,
-        PrePostStateRootVerifier<FullClient<RuntimeApi, ExecutorDispatch>, Block>,
+        VerifierClient<FullClient<RuntimeApi, ExecutorDispatch>, Block>,
         SystemDomainExtrinsicsBuilder<
             Block,
             FullClient<RuntimeApi, ExecutorDispatch>,
@@ -310,7 +310,7 @@ where
             client.clone(),
             executor.clone(),
             task_manager.spawn_handle(),
-            PrePostStateRootVerifier::new(client.clone()),
+            VerifierClient::new(client.clone()),
             SystemDomainExtrinsicsBuilder::new(client.clone(), Arc::new(executor)),
         ),
     ));
